Struct nepheliad::memory::area_frame_allocator::AreaFrameAllocator

pub struct AreaFrameAllocator {
    // some fields omitted
}

Represents an allocator which uses the Mulitboot2 memory map to tack available memory areas.

Methods

impl AreaFrameAllocator

fn new(kernel_start: usize, kernel_end: usize, multiboot_start: usize, multiboot_end: usize, memory_areas: MemoryAreaIter) -> AreaFrameAllocator

Constructs a new AreaFrameAllocator.

Parameters

kernel_start The address where the start of the kernel is loaded.

kernel_end The last address occupied by the kernel.

multiboot_start The address where the start of the multiboot info structure is loaded.

mulitboot_end The last address occupied by the multiboot info structure.

memory_areas An iterator over the multiboot memory map.

Returns

A newly initialized instance of AreaFrameAllocator