Module nepheliad::memory::paging

Defines the interface and types for memory paging.

Guarantees

A page table owns all of its subtables.

The recursively mapped P4 table is owned by a RecusivePageTable struct.

Structs

Entry
Represents an entry in the page table.

EntryFlags
The flags used to indicate the entry's state and contents.

Page
Represents a page of virtual memory.

RecusivePageTable
Represents a handle for the recursive page table hierarchy.

Constants

ACCESSED
Set by the CPU when the page is accessed.

DIRTY
Set by the CPU when a write to this page occurs.

GLOBAL
Indicates that this page is not flushed from cache on address space switch.

HUGE_PAGE
Indicates that this is a huge page.

NO_CACHE
Indicates that no cache (read or write) should be used for this page.

NO_EXECUTE
Indicates that code cannot be executed from this page.

PRESENT
Indicates the page this entry maps to is in memory.

USER_ACCESSIBLE
Indicates that user-mode processes can access this page.

WRITEABLE
Indicates that this page is writable.

WRITE_THROUGH
Indicates that writes should go directly to memory, bypassing the cache.

Functions

test_paging
Tests the paging Module.

Type Definitions

PhysicalAddress
Represents a physical memory address.

VirtualAddress
Represents a virtual memory address.