Struct nepheliad::memory::paging::RecusivePageTable

pub struct RecusivePageTable {
    // some fields omitted
}

Represents a handle for the recursive page table hierarchy.

Methods

impl RecusivePageTable

unsafe fn new() -> RecusivePageTable

Constructs a handle for the recursive page table hierarchy.

fn translate(&self, virtual_address: VirtualAddress) -> Option<PhysicalAddress>

Translates a virtual page address to the actual physical address.

Returns

The physical address mapped to the given virtual_address. Or None if the virtual_address is not mapped to any phsical address.

fn map_to<A>(&mut self, page: Page, frame: Frame, flags: EntryFlags, allocator: &mut A) where A: FrameAllocator

Maps a virtual memory page to a particular physical memory frame.

fn map<A>(&mut self, page: Page, flags: EntryFlags, allocator: &mut A) where A: FrameAllocator

Maps a virtual memory page to a physical memory frame.

This function will map a virtual memory page to an available physical memory frame. If no more frames are available it will panic.

Parameters

page The page to map to a frame.

flags The entry flags to apply to the new page entry.

allocator The allocator to use for allocating the fame to the page.

Panics

If there no available physical memory frames to allocate.

fn identity_map<A>(&mut self, frame: Frame, flags: EntryFlags, allocator: &mut A) where A: FrameAllocator

Remaps a particular frame to the associated page.